requirements.txt
Cython==0.24.1
Django==1.8
dill==0.2.5
djangorestframework==3.3.2
memory-profiler==0.39
numpy==1.10.1
pandas==0.19.0
psycopg2==2.6.2
pyfm==0.0.0
python-dateutil==2.5.3
pytz==2016.7
scikit-learn==0.18
scipy==0.18.1
six==1.10.0
sklearn==0.0
wsgiref==0.1.2
